# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Fishcake ecosystem - activity, drop, staking, mining and leaderboard",
        "title": "Fishcake Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/token/received/list": {
            "get": {
                "description": "Query token received records, includes Drop received, ActivityFinish refund and ERC20 Transfer received",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "summary": "Token received history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token contract address filter",
                        "name": "tokenType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Cursor: last record timestamp, default 0",
                        "name": "lastTimestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor: last record id",
                        "name": "lastId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default 10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "address": {
                                                "type": "string"
                                            },
                                            "amount": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "id": {
                                                "type": "string"
                                            },
                                            "log_index": {
                                                "type": "integer"
                                            },
                                            "timestamp": {
                                                "type": "integer"
                                            },
                                            "token_address": {
                                                "type": "string"
                                            },
                                            "tx_hash": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                },
                                "nextId": {
                                    "type": "string"
                                },
                                "nextTimestamp": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/token/sent/list": {
            "get": {
                "description": "Query token sent records, includes ActivityAdd sent and ERC20 Transfer sent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History"
                ],
                "summary": "Token sent history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token contract address filter",
                        "name": "tokenType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Cursor: last record timestamp, default 0",
                        "name": "lastTimestamp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor: last record id",
                        "name": "lastId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default 10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "address": {
                                                "type": "string"
                                            },
                                            "amount": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "id": {
                                                "type": "string"
                                            },
                                            "log_index": {
                                                "type": "integer"
                                            },
                                            "timestamp": {
                                                "type": "integer"
                                            },
                                            "token_address": {
                                                "type": "string"
                                            },
                                            "tx_hash": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                },
                                "nextId": {
                                    "type": "string"
                                },
                                "nextTimestamp": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}